find cube filter values
I wanted a way to determine where to set the maximum intensity value of the faces of the cube filter to find 'good' candles. I have been varying the number of z stacks in the images and thus have varied background in the top and bottom z slices. Therefore, I wanted to separate the values of the top and bottom faces of the cube filter from the side faces.